% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Hi how are you program that is well documented in roxygen}
\usage{
hi()
}
\value{
A character vector that say hi.
}
\description{
Hi how are you program that is well documented in roxygen
}
\examples{
hi()
}
